# build/resources/main/static/openapi-spec.yaml
openapi: 3.0.3
info:
  title: 캠퍼스 하우스 API
  description: 대학생 자취 커뮤니티 앱의 백엔드 API 문서
  version: 1.5.0
  contact:
    name: Campus House Team
    email: contact@campushouse.com
    url: https://campushouse.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: 개발 서버
  - url: https://api.campushouse.com
    description: 프로덕션 서버
security:
  - bearerAuth: []
paths:
...
  /api/buildings:
    get:
      tags:
        - 건물 정보
      summary: 모든 건물 조회
      description: 모든 건물 정보를 페이징으로 조회합니다.
      parameters:
        - name: page
          in: query
          description: 페이지 번호
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: 페이지당 건물 수
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          description: 정렬 기준
          schema:
            type: string
            default: "createdAt,desc"
      responses:
        '200':
          description: 건물 목록 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Building'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer

    post:
      tags:
        - 건물 정보
      summary: 건물 정보 생성
      description: 새로운 건물 정보를 생성합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBuildingRequest'
      responses:
        '200':
          description: 건물 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Building'
        '400':
          description: 잘못된 요청
  /api/buildings/{buildingId}:
    get:
      tags:
        - 건물 정보
      summary: 건물 상세 조회
      description: 특정 건물의 상세 정보를 조회합니다.
      parameters:
        - name: buildingId
          in: path
          required: true
          description: 건물 ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: 건물 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Building'
        '404':
          description: 건물을 찾을 수 없음
  /api/buildings/{buildingId}/building-info:
    get:
      tags:
        - 건물 정보
      summary: 건물 기본 정보 조회
      description: 특정 건물의 기본 정보를 조회합니다.
      parameters:
        - name: buildingId
          in: path
          required: true
          description: 건물 ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: 건물 기본 정보 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Building'
        '404':
          description: 건물을 찾을 수 없음
  /api/buildings/{buildingId}/reviews:
    get:
      tags:
        - 건물 정보
      summary: 건물 후기 목록 조회
      description: 특정 건물의 실거주자 후기 목록을 조회합니다.
      parameters:
        - name: buildingId
          in: path
          required: true
          description: 건물 ID
          schema:
            type: integer
            example: 1
        - name: page
          in: query
          description: 페이지 번호
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: 페이지당 후기 수
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 건물 후기 목록 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/BuildingReview'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
  /api/buildings/{buildingId}/qnas:
    get:
      tags:
        - 건물 정보
      summary: 건물 질문 목록 조회
      description: 특정 건물의 질문 목록을 조회합니다. (Post 엔티티의 QUESTION 게시판 기반)
      parameters:
        - name: buildingId
          in: path
          required: true
          description: 건물 ID
          schema:
            type: integer
            example: 1
        - name: page
          in: query
          description: 페이지 번호
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: 페이지당 질문 수
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 건물 질문 목록 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
  /api/buildings/{buildingId}/questions:
    post:
      tags:
        - 건물 정보
      summary: 건물별 질문 작성
      description: 특정 건물에 대한 질문을 작성합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: buildingId
          in: path
          required: true
          description: 건물 ID
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 질문 제목
                  example: "이 건물 주차장은 어떻게 되나요?"
                content:
                  type: string
                  description: 질문 내용
                  example: "주차장이 있는지, 월 주차비는 얼마인지 궁금합니다."
      responses:
        '200':
          description: 질문 작성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: 잘못된 요청
        '401':
          description: 인증 필요
  /api/buildings/{buildingId}/transfers:
    get:
      tags:
        - 건물 정보
      summary: 건물 양도 목록 조회
      description: 특정 건물의 양도 관련 글 목록을 조회합니다. (Post 엔티티의 TRANSFER 게시판 기반)
      parameters:
        - name: buildingId
          in: path
          required: true
          description: 건물 ID
          schema:
            type: integer
            example: 1
        - name: page
          in: query
          description: 페이지 번호
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: 페이지당 양도 글 수
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 건물 양도 목록 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer

    post:
      tags:
        - 건물 정보
      summary: 건물별 양도 글 작성
      description: 특정 건물에 대한 양도 글을 작성합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: buildingId
          in: path
          required: true
          description: 건물 ID
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 양도 글 제목
                  example: "캠퍼스 하우스 A동 양도합니다"
                content:
                  type: string
                  description: 양도 글 내용
                  example: "양도비 50만원에 양도합니다. 연락주세요."
      responses:
        '200':
          description: 양도 글 작성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: 잘못된 요청
        '401':
          description: 인증 필요

    put:
      tags:
        - 건물 정보
      summary: 건물 정보 수정
      description: 건물 정보를 수정합니다.
      parameters:
        - name: buildingId
          in: path
          required: true
          description: 건물 ID
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBuildingRequest'
      responses:
        '200':
          description: 건물 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Building'
        '404':
          description: 건물을 찾을 수 없음

    delete:
      tags:
        - 건물 정보
      summary: 건물 정보 삭제
      description: 건물 정보를 삭제합니다.
      parameters:
        - name: buildingId
          in: path
          required: true
          description: 건물 ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: 건물 삭제 성공
        '404':
          description: 건물을 찾을 수 없음
  /api/buildings/search:
    get:
      tags:
        - 건물 정보
      summary: 키워드로 건물 검색
      description: 키워드로 건물을 검색합니다.
      parameters:
        - name: keyword
          in: query
          required: true
          description: 검색 키워드
          schema:
            type: string
            example: "서울대학교"
        - name: page
          in: query
          description: 페이지 번호
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: 페이지당 건물 수
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 건물 검색 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Building'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
  /api/buildings/search/filters:
    get:
      tags:
        - 건물 정보
      summary: 건물 정보 필터링
      description: 다양한 조건으로 건물을 필터링합니다.
      parameters:
        - name: minDeposit
          in: query
          description: 최소 보증금 (만원)
          schema:
            type: number
            format: decimal
            example: 500
        - name: maxDeposit
          in: query
          description: 최대 보증금 (만원)
          schema:
            type: number
            format: decimal
            example: 2000
        - name: minMonthlyRent
          in: query
          description: 최소 월세 (만원)
          schema:
            type: number
            format: decimal
            example: 30
        - name: maxMonthlyRent
          in: query
          description: 최대 월세 (만원)
          schema:
            type: number
            format: decimal
            example: 80
        - name: minJeonse
          in: query
          description: 최소 전세 (만원)
          schema:
            type: number
            format: decimal
            example: 5000
        - name: maxJeonse
          in: query
          description: 최대 전세 (만원)
          schema:
            type: number
            format: decimal
            example: 30000
        - name: parkingRequired
          in: query
          description: 주차장 필요 여부
          schema:
            type: boolean
            example: true
        - name: elevatorRequired
          in: query
          description: 엘리베이터 필요 여부
          schema:
            type: boolean
            example: true
        - name: maxWalkingTime
          in: query
          description: 최대 도보 시간 (분)
          schema:
            type: integer
            example: 10
        - name: page
          in: query
          description: 페이지 번호
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: 페이지당 건물 수
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 필터링 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Building'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
...
  /api/buildings/recent:
    get:
      tags:
        - 건물 정보
      summary: 최근 등록된 건물 조회
      description: 최근에 등록된 건물 목록을 조회합니다.
      parameters:
        - name: page
          in: query
          description: 페이지 번호
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: 페이지당 건물 수
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 최근 건물 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Building'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT 토큰을 입력하세요. 예: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
...